resolve pr errors
