+++ openbas-api/src/main/java/io/openbas/executors/Injector.java
@@ -41,7 +41,7 @@ public abstract class Injector {
   public abstract ExecutionProcess process(Execution execution, ExecutableInject injection)
       throws Exception;
 
-  public PayloadOutput getPayloadOutput(String externalId) {
+  public StatusPayload getPayloadOutput(String externalId) {
     return null;
   }
