+++ openbas-api/src/main/java/io/openbas/migration/V3_49__Update_Commands_In_Inject_Status.java
@@ -3,7 +3,7 @@ package io.openbas.migration;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.openbas.database.model.PayloadCommandBlock;
-import io.openbas.database.model.PayloadOutput;
+import io.openbas.database.model.StatusPayload;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
@@ -54,8 +54,8 @@ public class V3_49__Update_Commands_In_Inject_Status extends BaseJavaMigration {
           }
         }
         String externalId = jsonNode.get("external_id").asText();
-        PayloadOutput payloadOutput =
-            new PayloadOutput(
+        StatusPayload statusPayload =
+            new StatusPayload(
                 null,
                 null,
                 null,
@@ -69,7 +69,7 @@ public class V3_49__Update_Commands_In_Inject_Status extends BaseJavaMigration {
                 null,
                 List.of(new PayloadCommandBlock(null, contentString, cleanupCommandList)),
                 null);
-        String value = mapper.writeValueAsString(payloadOutput);
+        String value = mapper.writeValueAsString(statusPayload);
         statement.setString(1, value);
         statement.setString(2, statusId);
         statement.addBatch();
