+++ openbas-api/src/main/java/io/openbas/rest/atomic_testing/AtomicTestingApi.java
@@ -47,7 +47,7 @@ public class AtomicTestingApi extends RestBehavior {
   @LogExecutionTime
   @GetMapping("/{injectId}/payload")
   @Tracing(name = "Get the payload of an atomic testing", layer = "api", operation = "GET")
-  public PayloadOutputDto findAtomicTestingPayload(@PathVariable String injectId) {
+  public StatusPayloadOutput findAtomicTestingPayload(@PathVariable String injectId) {
     return atomicTestingService.findPayloadOutputByInjectId(injectId);
   }
